update project to zyspotify, use sqlite3 db instead of json file archive (#1)
These changes were the main reasons for the fork. The most notable is using an sqlite3 db for keeping track of what has been downloaded instead of a json file. This is required as the file is not scalable to mass downloading (>10k songs). It also offers a great speedup, caching as much as possible so that if the program is stopped, it can be resumed in a matter of seconds. This is compared to the upstream where too many requests for information were taking place that only needed to happen once.

Yes, this breaks compatibility with the old archive.json file, that is intentional. No support is planned for migration.

Also includes the first batch of updates for the new project name, README changes, etc.